refactor minikube install script

+++ minikube/install.sh
@@ -2,8 +2,13 @@
 
 case $(uname -s) in
   'Linux')
-    curl -LO https://storage.googleapis.com/minikube/releases/latest/minikube-linux-amd64
-    sudo install minikube-linux-amd64 /usr/local/bin/minikube
+    if [ -f /usr/local/bin/minikube ]; then
+      echo "minikube is already installed. Skipping..."
+    else
+      curl -LO https://storage.googleapis.com/minikube/releases/latest/minikube-linux-amd64
+      sudo install minikube-linux-amd64 /usr/local/bin/minikube
+      rm -rf minikube-linux-amd64
+    fi
     ;;
   'Darwin') 
     brew install minikube
